requirements.txt
numpy==1.9.1
nltk==3.0.1
requests==2.2.1
nose==1.3.4
xdiagnose==3.6.3build2
beautifulsoup4==4.3.2
pyzmq==14.5.0
pyserial==2.7
